fix(rn) improve rnsdk version script
Make sure alll dependencies are updated when we run it.
